requirements.txt
# python > 3.8
python-telegram-bot
asyncio
python-dotenv
requests